backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.0.0
python-multipart==0.0.6
spacy==3.7.2
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl
python-dotenv==1.0.0
openai>=1.2.0
textstat==0.7.3
requests
beautifulsoup4
google-api-python-client
anthropic
tiktoken 